requirements.txt
alpaca-py==0.42.1
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cramjam==2.11.0
curl_cffi==0.13.0
fastapi==0.116.2
fastparquet==2024.11.0
frozendict==2.4.6
fsspec==2025.9.0
idna==3.10
msgpack==1.1.1
multitasking==0.0.12
mypy==1.11.2
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sseclient-py==1.8.0
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
